music_connect/connectors/spotify_connector/requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
pydantic-settings
python-dotenv
redis
psycopg2-binary